requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.3
Flask-Cors==5.0.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.4.1
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
idna==3.8
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==1.26.20
Werkzeug==3.0.4
zipp==3.20.2
